requirements.txt
arrow==1.3.0
sqlalchemy==1.4.50
sqlalchemy-utils==0.41.1

# tests
factory-boy==3.3.0
faker==20.1.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-env==1.1.3
pytest-factoryboy==2.6.0
pytest-mock==3.12.0
pytest-ordering==0.6
pytest-sugar==0.9.7

# tools
autotyping==23.3.0
black==23.11.0
bump2version==1.0.1
flake8==6.1.0
greenlet==3.0.1
identify==2.5.32
interrogate==1.5.0
ipython==8.18.1
isort==5.12.0
mccabe==0.7.0
pipdeptree==2.13.1
pipfile-requirements==0.3.0
pre-commit==3.5.0
pylint==2.17.7
pylint-pytest==1.1.6
